Drop pal_asn1_print in favor of the managed code that is already used on macOS.
